Added Django's pagination to our blog
In this we added pagination capabilities to our blog using Django's pagination library.
We created a hew pagination html which would handle all the pagination in the blogs template folder and our blog app.
Modified the views.py file in our blog app to add pagination and updated list.html file to have pagination added to it.